Add support for array values in where conditions
This adds the possibility to have arrays as values in data.
So an attribute can have a value of `[1,2,3]` for example and you’ll be
able to search for it by doing `.where(attribute: 3)` or
`.where(attribute: [1,3])`. Other values are still handled correctly, in
particular `nil`. So if you have data with a `nil` attribute, it will
correctly returns the records too: `.where(attribute: nil)`.