Add thread safety to ActiveFile
ActiveFile is not thread-safe, which results in the following: if the
very first request to ActiveFile goes simultaneously from even two
threads, the first one is going to return as if the data was empty. The
second request will immediately fix that, but that unlucky thread will
receive a bad value. It may store it in the database or memoize it,
which will result in further problem escalation.

We can't just synchronize it, because we check the variable very far
from reload. While we must synchronize, it's not enough. We have to flag
it as 'done' only after it's loaded.

Another problem is that we touch `all` in process of reloading, which
will check `data_loaded` flag and try to reload again. To counter that,
I come out with a method that won't trigger reload, to use it in methods
that work during reload.